Fix bugs in savedata manager

+++ Common/GPU/Vulkan/thin3d_vulkan.cpp
@@ -271,7 +271,7 @@ public:
 	}
 
 	void SetDynamicUniformData(const void *data, size_t size) {
-		_dbg_assert_(size <= uboSize_);
+		_dbg_assert_((int)size <= uboSize_);
 		memcpy(ubo_, data, size);
 	}
 
@@ -822,7 +822,7 @@ bool VKTexture::Create(VkCommandBuffer cmd, VulkanBarrierBatch *postBarriers, Vu
 	if (desc.initData.size()) {
 		UpdateInternal(cmd, pushBuffer, desc.initData.data(), desc.initDataCallback, (int)desc.initData.size());
 		// Generate the rest of the mips automatically.
-		if (desc.initData.size() < mipLevels_) {
+		if ((int)desc.initData.size() < mipLevels_) {
 			vkTex_->GenerateMips(cmd, (int)desc.initData.size(), false);
 			layout = VK_IMAGE_LAYOUT_GENERAL;
 		}
